Add files via upload
Current version of the ACARs project. As of right now I only have the GUI part set up. The next step is to figure out how to merge all the pages. I.e. when the next button is clicked, instead of the next page as a individual pop-up, it shows up as the next page in the same window as the previous page.